+++ packages/tools/src/tools/ScaleOverlayTool.ts
@@ -139,7 +139,10 @@ class ScaleOverlayTool extends AnnotationDisplayTool {
 
       addAnnotation(newAnnotation, viewport.element);
       annotation = newAnnotation;
-    } else if (this.editData.annotation && this.editData.annotation.data.viewportId == viewport.id) {
+    } else if (
+      this.editData.annotation &&
+      this.editData.annotation.data.viewportId == viewport.id
+    ) {
       this.editData.annotation.data.handles.points =
         viewportCanvasCornersInWorld;
       this.editData.annotation.data.viewportId = viewport.id;