+++ packages/tools/src/tools/annotation/PlanarFreehandContourSegmentationTool.ts
@@ -1,8 +1,6 @@
-import type { SplineContourSegmentationAnnotation } from '../../types/ToolSpecificAnnotationTypes';
 import { utilities } from '@cornerstonejs/core';
-import type { PublicToolProps, AnnotationRenderContext } from '../../types';
+import type { PublicToolProps } from '../../types';
 import PlanarFreehandROITool from './PlanarFreehandROITool';
-import { triggerSegmentationDataModified } from '../../stateManagement/segmentation/triggerSegmentationEvents';
 
 class PlanarFreehandContourSegmentationTool extends PlanarFreehandROITool {
   static toolName;
@@ -29,25 +27,6 @@ class PlanarFreehandContourSegmentationTool extends PlanarFreehandROITool {
   protected isContourSegmentationTool(): boolean {
     // Re-enable contour segmentation behavior disabled by PlanarFreehandROITool
     return true;
-  }
-
-  protected renderAnnotationInstance(
-    renderContext: AnnotationRenderContext
-  ): boolean {
-    const { annotation } = renderContext;
-    const { invalidated } = annotation;
-    const renderResult = super.renderAnnotationInstance(renderContext);
-
-    // Trigger the event only for invalid annotations
-    if (invalidated) {
-      const { segmentationId } = (<SplineContourSegmentationAnnotation>(
-        annotation
-      )).data.segmentation;
-
-      triggerSegmentationDataModified(segmentationId);
-    }
-
-    return renderResult;
   }
 }