+++ packages/tools/src/tools/segmentation/strategies/BrushStrategy.ts
@@ -196,20 +196,23 @@ export default class BrushStrategy {
       return operationData.preview;
     }
 
-    if (isVolumeSegmentation(operationData)) {
+    if (isVolumeSegmentation(operationData, viewport)) {
       const { referencedVolumeId, volumeId } =
         operationData as LabelmapToolOperationDataVolume;
 
-      const imageVolume = cache.getVolume(referencedVolumeId);
       const segmentation = cache.getVolume(volumeId);
 
-      if (
-        !csUtils.isEqual(segmentation.dimensions, imageVolume.dimensions) ||
-        !csUtils.isEqual(segmentation.direction, imageVolume.direction)
-      ) {
-        throw new Error(
-          'Only source data the same dimensions/size/orientation as the segmentation currently supported.'
-        );
+      if (referencedVolumeId) {
+        const imageVolume = cache.getVolume(referencedVolumeId);
+
+        if (
+          !csUtils.isEqual(segmentation.dimensions, imageVolume.dimensions) ||
+          !csUtils.isEqual(segmentation.direction, imageVolume.direction)
+        ) {
+          throw new Error(
+            'Only source data the same dimensions/size/orientation as the segmentation currently supported.'
+          );
+        }
       }
     }