+++ packages/tools/src/tools/segmentation/strategies/utils/getStrategyData.ts
@@ -7,21 +7,28 @@ const { VoxelManager } = utilities;
 
 function getStrategyData({ operationData, viewport }) {
   let segmentationImageData, segmentationScalarData, imageScalarData;
-  let dimensions: Types.Point3;
-  if (isVolumeSegmentation(operationData)) {
+  let imageDimensions: Types.Point3;
+  let segmentationDimensions: Types.Point3;
+  if (isVolumeSegmentation(operationData, viewport)) {
     const { volumeId, referencedVolumeId } = operationData;
 
     const segmentationVolume = cache.getVolume(volumeId);
-    const imageVolume = cache.getVolume(referencedVolumeId);
 
-    if (!segmentationVolume || !imageVolume) {
+    if (!segmentationVolume) {
       return;
+    }
+
+    // we only need the referenceVolumeId if we do thresholding
+    // but for other operations we don't need it so make it optional
+    if (referencedVolumeId) {
+      const imageVolume = cache.getVolume(referencedVolumeId);
+      imageScalarData = imageVolume.getScalarData();
+      imageDimensions = imageVolume.dimensions;
     }
 
     ({ imageData: segmentationImageData } = segmentationVolume);
     segmentationScalarData = segmentationVolume.getScalarData();
-    imageScalarData = imageVolume.getScalarData();
-    dimensions = imageVolume.dimensions;
+    segmentationDimensions = segmentationVolume.dimensions;
   } else {
     const { imageIdReferenceMap, segmentationRepresentationUID } =
       operationData as LabelmapToolOperationDataStack;
@@ -51,21 +58,29 @@ function getStrategyData({ operationData, viewport }) {
     // This is the pixel data of the image that is being segmented in the cache
     // and we need to use this to for the modification
     imageScalarData = image.getPixelData();
-    dimensions = [image.columns, image.rows, 1];
+    imageDimensions = [image.columns, image.rows, 1];
+    segmentationDimensions = [
+      segmentationImage.columns,
+      segmentationImage.rows,
+      1,
+    ];
   }
+
+  const segmentationVoxelManager = VoxelManager.createVolumeVoxelManager(
+    segmentationDimensions,
+    segmentationScalarData
+  );
+
+  const imageVoxelManager =
+    imageDimensions &&
+    VoxelManager.createVolumeVoxelManager(imageDimensions, imageScalarData);
 
   return {
     segmentationImageData,
     segmentationScalarData,
-    segmentationVoxelManager: VoxelManager.createVolumeVoxelManager(
-      dimensions,
-      segmentationScalarData
-    ),
+    segmentationVoxelManager,
     imageScalarData,
-    imageVoxelManager: VoxelManager.createVolumeVoxelManager(
-      dimensions,
-      imageScalarData
-    ),
+    imageVoxelManager,
   };
 }